removes the gui filters flag and makes gui filters the default
